refactor: navbar links cleanup
Seperated NavbarDropdown and NavbarDropdownItem into separate components.
Made the whole navbar generation to be based of a links list.
